% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes-result.R
\name{validate_bayes_result}
\alias{validate_bayes_result}
\title{Validate a bayes_result}
\usage{
validate_bayes_result(x)
}
\arguments{
\item{x}{A \code{bayes_result}.}
}
\value{
\code{x} invisibly on success.
}
\description{
Validate a bayes_result
}
